Break down the replace/module before adding to it
