+++ src/replace/module/index.browser.js
